+++ examples/out/rfc/node-interop/multi-package-dev/main.go
@@ -4,8 +4,8 @@ import "os"
 import fmt "fmt"
 
 func main() {
-	if os.Getenv("FORST_INVOKE_ONLY") == "1" {
-		println("multipackage-dev invoke-only")
+	if os.Getenv("FORST_SKIP_NODE_HOST") == "1" {
+		println("multipackage-dev skip-node-host")
 	} else {
 		ready, readyErr := forst_node_callsync_host_ts_hostPing()
 		if !(readyErr == nil) {